backend/web/core/requirements/base.txt
django==3.2.2

# for PostgreSql database
psycopg2==2.8.6

# Ddos defender: required redis
django-defender==0.9.0

# redis
redis==3.5.3

# for RestAPI
djangorestframework==3.12.4
drf-yasg==1.20.0 # Swagger

# for auth system
django-allauth==0.42.0
dj-rest-auth==2.1.3
djangorestframework-simplejwt==4.6.0

# For backend working
celery==5.0.5

# for frontend
django-cors-headers==3.7.0

# API keys
djangorestframework-api-key==2.0.0

# For testing
flake8==3.9.0

# Async Support
uvicorn==0.13.3
httpx==0.16.1

# Cashe Support
django-cacheops==5.1

# ImageField
pillow==8.1.1

# JsonLogger
python-json-logger==2.0.1

# Translation
django-rosetta==0.9.5

# Jaeger OpenTracing
jaeger-client==4.4.0
django_opentracing==1.1.0

# Monitoring System
sentry-sdk==1.0.0

# MicroService Requests
django-microservice-request>=0.1.3.3

# Payments
liqpay-python @ git+https://github.com/liqpay/sdk-python@f3a81847f2b518c3936461880e0451eb02f33b25
